requirements.txt
natsort==8.4.0
Pillow==11.0.0
